+++ src/containers/main-page/VideoPlayer/Name.tsx
@@ -7,7 +7,10 @@ import { useConnection } from 'common/providers/hooks/useConnection';
 const Name = function (): JSX.Element {
   const { name } = useConnection();
   return (
-    <Typography variant="h5" gutterBottom>{name || 'Name'}</Typography>
+    <Typography variant="h5" gutterBottom>
+      {'Me: '}
+      { name || ' Name'}
+    </Typography>
   );
 };
